test: classify broadcast contention proof
